requirements.txt
amqp==5.1.1
asgiref==3.6.0
billiard==3.6.4.0
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
contourpy==1.0.7
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
cycler==0.11.0
DAWG-Python==0.7.2
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.2.1
django-allauth==0.54.0
django-prometheus==2.3.1
django-rest-auth==0.9.5
django-rest-auth-forked==0.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
docx==0.2.4
drf-yasg==1.21.5
et-xmlfile==1.1.0
fonttools==4.39.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
joblib==1.2.0
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.7.1
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openapi-codec==1.3.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.1
pdfminer.six==20221105
Pillow==9.5.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
pycparser==2.21
PyJWT==2.6.0
pymorphy2==0.9.1
pymorphy2-dicts-ru==2.4.417127.4579844
pymorphy3==1.2.0
pymorphy3-dicts-ru==2.4.417150.4580142
pymystem3==0.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-docx==0.8.11
python3-openid==3.2.0
pytz==2023.3
redis==4.5.5
regex==2023.5.5
requests==2.30.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.24
ruamel.yaml.clib==0.2.7
seaborn==0.12.2
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
tqdm==4.65.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
wcwidth==0.2.6
